Workaround jruby-openssl bug
When creating a CRL from scratch, the internal state of the CRL is
incorrect when calling #to_pem, resulting in an invalid CRL.  A bunch of
tests are failing because of this.

Implement a workaround by serializing the CRL to der format and use that
to initialize a new CRL object.