Fix updating extentions in CRL
Updating extensions in-place does not work with all Ruby
implementations.  Moreover, building a `crlNumber` extension does not
work the same way on MRI and jruby, its bugs all the way down :-(.

The first issue was already encountered with MRI:
https://github.com/smortex/puppet-renew-certificate/blob/761d5e768933aae0233e77aeac4aea01d3fd2fa8/exe/puppet-renew-certificate#L125

So, play safe: remove extensions and re-create them one by one instead
of trying to update them in-place.  Also check what implementation of
Ruby we are running to add the `crlNumber` as pleased the current
ruby virtual machine.