Remove jruby unit testing
ruby/openssl and jruby/jruby-openssl have slightly different
interfaces, and supporting both would require more adjustments, but as
the openvoxservec ca command run in the context of MRI, we do not depend
on this work for now.

This commit can be reverted if we want to run the code on jruby.